\name{aurelhy-package}
\alias{aurelhy-package}

\docType{package}

\title{Hydrometeorological Interpolation}

\description{Hydrometeorological interpolation using the AURELHY method.}

\details{
\tabular{ll}{Package: \tab aurelhy\cr
Title: \tab Hydrometeorological Interpolation\cr
Type: \tab Package\cr
Version: \tab 1.0-7\cr
Date: \tab 2015-07-16\cr
Authors@R: \tab c(person("Philippe", "Grosjean", role = c("aut", "cre"),
  email = "phgrosjean@sciviews.org"))\cr
URL: \tab https://github.com/phgrosjean/aurelhy\cr
Depends: \tab sp, gstat\cr
Suggests: \tab shapefiles\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Encoding: \tab UTF-8\cr
Author: Philippe Grosjean [aut, cre]\cr
Maintainer: \tab Philippe Grosjean <phgrosjean@sciviews.org>\cr}
}

\author{Philippe Grosjean <phgrosjean@sciviews.org>}

\keyword{ package }
