\name{mbord}
\Rdversion{1.1}
\alias{mbord}
\non_function{}
\title{ A geoshapes object with a polygon of the Morocco borders }
\usage{data(mbord)}
\description{
The \code{mbord} dataset is a 'geoshapes' object (georeferenced shapes).
}
\format{
  This 'geoshapes' object contains a polygon defining the borders of Morocco in
  decimal degrees (longitudes - latitudes). 
}
\source{
TODO: add source of these data.
}

\seealso{ \code{\link{geoshapes}}, \code{\link{morocco}} , \code{\link{mmask}}}

\keyword{datasets}
