\name{morocco}
\Rdversion{1.1}
\alias{morocco}
\non_function{}
\title{ A digital elevation model with a grid of roughly 1km x 1km of Morocco }
\usage{data(morocco)}
\description{
The \code{morocco} dataset is a 'dem' object.
}
\format{
  This 'dem' object contains a grid of 1986x1866 elevetion points in m. 
}
\source{
TODO: add source of these data.
}

\seealso{ \code{\link{dem}} }

\keyword{datasets}
