\name{mpet}
\Rdversion{1.1}
\alias{mpet}
\non_function{}
\title{ A geopoints object with PET values measured at different weather stations }
\usage{data(mpet)}
\description{
The \code{mpet} dataset is a 'geopoints' object (georeferenced points data).
}
\format{
  This 'geopoints' object contains a data frame with PET measurements (\code{D1} - \code{D36})
  made at 18 weather stations (names in \code{STATION} and coordinates in \code{x} and \code{y}). 
}

\seealso{ \code{\link{geopoints}}, \code{\link{morocco}},
  \code{\link{mbord}}, \code{\link{mrain}}}

\keyword{datasets}
