\name{mrain}
\Rdversion{1.1}
\alias{mrain}
\non_function{}
\title{ A geopoints object with normalized rain values (mm) measured
  at different weather stations }
\usage{data(mrain)}
\description{
The \code{mrain} dataset is a 'geopoints' object (georeferenced points data).
}
\format{
  This 'geopoints' object contains a data frame with rain measurements (\code{rain} variable)
  made at 43 weather stations (coordinates in \code{x} and \code{y} variables). 
}

\seealso{ \code{\link{geopoints}}, \code{\link{morocco}},
  \code{\link{mbord}}, \code{\link{mpet}}}

\keyword{datasets}
