\name{mseadist}
\alias{mseadist}
\non_function{}
\title{ A geomat object with distance from the sea for the morocco data }
\usage{data(mseadist)}
\description{
The \code{mseadist} dataset is a 'geomat' object (georeferenced data).
}
\format{
  This 'geomat' object contains a matrix with distance from the sea for all
  points in the \code{morocco} dataset. It can be used as supplemntary variable
  for AURELHY interpolation.
}

\seealso{ \code{\link{geomat}}, \code{\link{morocco}},
  \code{\link{mbord}}, \code{\link{mpet}}}

\keyword{datasets}
