\name{mbord}
\alias{mbord}
\non_function{}
\title{ A geoshapes object with a polygon of the area to analyze (around Morocco) }
\usage{data(mbord)}
\description{
The \code{mbord} dataset is a 'geoshapes' object (georeferenced shapes).
}
\format{
  This 'geoshapes' object contains a polygon defining the area to analyze around
  Morocco in decimal degrees (longitudes - latitudes). 
}

\seealso{ \code{\link{geoshapes}}, \code{\link{morocco}} , \code{\link{mmask}}}

\keyword{datasets}
