\name{morocco}
\alias{morocco}
\non_function{}
\title{ A digital elevation model with a grid of roughly 1km x 1km of Morocco }
\usage{data(morocco)}
\description{
The \code{morocco} dataset is a 'geotm' object (georeferenced terrain model).
}
\format{
  This 'geotm' object contains a grid of 1986x1866 elevation points (in m). 
}

\seealso{ \code{\link{geotm}}, \code{\link{mbord}} , \code{\link{mmask}}}

\keyword{datasets}
